refactor(i18n): refactor to support crowdin
