+++ src/i18n/strings/es.json
