+++ src/i18n/strings/fr.json
