+++ src/i18n/strings/zh-Hans.json
